Added Line chart component
